+++ index.js
@@ -12,21 +12,32 @@ class FakeAmqpError extends Error {
   }
 }
 
+class FakeAmqpNotFoundError extends FakeAmqpError {
+  constructor(type, name, killConnection = false) {
+    super(`Channel closed by server: 404 (NOT-FOUND) with message "NOT_FOUND - no ${type} '${name}' in vhost '/'`, 404, true, killConnection);
+  }
+}
+
 const connections = [];
 
-module.exports = Fake();
+module.exports = Fake('3.5');
 
-function Fake() {
+function Fake(minorVersion) {
+  let defaultVersion = Number(minorVersion);
   return {
     connections,
     resetMock,
+    setVersion(minor) {
+      const n = Number(minor);
+      if (!isNaN(n)) defaultVersion = n;
+    },
     connect,
   };
 
   function connect(amqpUrl, ...args) {
     const {_broker} = connections.find((conn) => conn.options[0] === amqpUrl) || {};
     const broker = _broker || Broker();
-    const connection = Connection(broker, amqpUrl, ...args);
+    const connection = Connection(broker, defaultVersion, amqpUrl, ...args);
     connections.push(connection);
     return resolveOrCallback(args.slice(-1)[0], null, connection);
   }
@@ -37,7 +48,7 @@ function Fake() {
     }
   }
 
-  function Connection(broker, ...connArgs) {
+  function Connection(broker, version, ...connArgs) {
     const emitter = new EventEmitter();
     const options = connArgs.filter((a) => typeof a !== 'function');
     let closed = false;
@@ -46,6 +57,7 @@ function Fake() {
     return {
       _id: generateId(),
       _broker: broker,
+      _version: version,
       get _closed() {
         return closed;
       },
@@ -74,6 +86,8 @@ function Fake() {
         if (idx > -1) connections.splice(idx, 1);
         channels.splice(0).forEach((channel) => channel.close());
 
+        emitter.emit('close');
+
         return resolveOrCallback(args.slice(-1)[0]);
       },
       on(...args) {
@@ -93,13 +107,18 @@ function Fake() {
     let closed = false, prefetch = 10000;
     const emitter = new EventEmitter();
     const channelName = 'channel-' + generateId();
+    const version = connection._version;
 
     broker.on('return', emitReturn);
 
     return {
       _broker: broker,
+      _version: version,
       get _emitter() {
         return emitter;
+      },
+      get _closed() {
+        return closed;
       },
       assertExchange(...args) {
         return callBroker(assertExchange, ...args);
@@ -125,17 +144,21 @@ function Fake() {
           };
         }
       },
-      bindExchange(...args) {
-        return callBroker(broker.bindExchange, ...args);
+      bindExchange(destination, source, ...args) {
+        return Promise.all([this.checkExchange(source), this.checkExchange(destination)]).then(() => {
+          return callBroker(broker.bindExchange, source, destination, ...args);
+        });
       },
-      bindQueue(...args) {
-        return callBroker(broker.bindQueue, ...args);
+      bindQueue(queue, source, ...args) {
+        return Promise.all([this.checkQueue(queue), this.checkExchange(source)]).then(() => {
+          return callBroker(broker.bindQueue, queue, source, ...args);
+        });
       },
       checkExchange(name, ...args) {
         return callBroker(check, ...args);
 
         function check() {
-          if (!broker.getExchange(name)) throw new FakeAmqpError(`exchange with name ${name} not found`, 404);
+          if (!broker.getExchange(name)) throw new FakeAmqpError(`Channel closed by server: 404 (NOT-FOUND) with message "NOT_FOUND - no exchange '${name}' in vhost '/'`, 404, true);
           return true;
         }
       },
@@ -145,7 +168,7 @@ function Fake() {
         function check() {
           let queue;
           if (!(queue = broker.getQueue(name))) {
-            throw new FakeAmqpError(`Channel closed by server: 404 (NOT-FOUND) with message "NOT_FOUND - no queue '${name}' in vhost '/'`, 404, true);
+            throw new FakeAmqpNotFoundError('queue', name);
           }
 
           return {
@@ -159,15 +182,27 @@ function Fake() {
 
         function getMessage(...getargs) {
           const q = broker.getQueue(queue);
-          if (!q) throw new FakeAmqpError(`queue with name ${queue} not found`, 404);
+          if (!q) throw new FakeAmqpNotFoundError('queue');
           return q.get(...getargs) || false;
         }
       },
-      deleteExchange(...args) {
-        return callBroker(broker.deleteExchange, ...args);
+      deleteExchange(exchange, ...args) {
+        return callBroker(check, ...args);
+
+        function check() {
+          const result = broker.deleteExchange(exchange, ...args);
+          if (!result && version < 3.2) throw new FakeAmqpNotFoundError('exchange', exchange);
+          return result;
+        }
       },
-      deleteQueue(...args) {
-        return callBroker(broker.deleteQueue, ...args);
+      deleteQueue(queue, ...args) {
+        return callBroker(check, ...args);
+
+        function check() {
+          const result = broker.deleteQueue(queue, ...args);
+          if (!result && version < 3.2) throw new FakeAmqpNotFoundError('queue', queue);
+          return result;
+        }
       },
       publish(exchange, routingKey, content, options, callback) {
         if (!Buffer.isBuffer(content)) throw new TypeError('content is not a buffer');
@@ -175,34 +210,77 @@ function Fake() {
           options = {...options, confirm: makeConfirmCallback(callback)};
         }
 
-        callBroker(broker.publish, exchange, routingKey, content, options);
+        this.checkExchange(exchange).then(() => {
+          return callBroker(broker.publish, exchange, routingKey, content, options);
+        }).catch((err) => {
+          emitter.emit('error', err);
+        });
+
         return true;
       },
-      purgeQueue(...args) {
-        return callBroker(broker.purgeQueue, ...args);
+      purgeQueue(queue, ...args) {
+        return callBroker(check, ...args);
+
+        function check() {
+          const result = broker.purgeQueue(queue);
+          if (!result && version < 3.2) throw new FakeAmqpNotFoundError('queue', queue);
+          return result === undefined ? undefined : {messageCount: result};
+        }
       },
       sendToQueue(queue, content, options, callback) {
         if (!Buffer.isBuffer(content)) throw new TypeError('content is not a buffer');
-        return confirmChannel ? sendToQueue() : callBroker(broker.sendToQueue, queue, content, options);
-
-        function sendToQueue() {
-          const confirm = makeConfirmCallback(callback);
-          broker.sendToQueue(queue, content, {...options, confirm});
+        if (confirmChannel) {
+          options = {...options, confirm: makeConfirmCallback(callback)};
         }
+
+        this.checkQueue(queue).then(() => {
+          return callBroker(broker.sendToQueue, queue, content, options);
+        }).catch((err) => {
+          emitter.emit('error', err);
+        });
+
+        return true;
       },
-      unbindExchange(...args) {
-        return callBroker(broker.unbindExchange, ...args);
-      },
-      unbindQueue(...args) {
-        return callBroker(broker.unbindQueue, ...args);
-      },
-      consume(queue, onMessage, options = {}, callback) {
-        return callBroker(check, queue, onMessage, options, callback);
+      unbindExchange(destination, source, pattern, ...args) {
+        return callBroker(check, ...args);
 
         function check() {
+          const q = broker.getExchange(destination);
+          if (!q) throw new FakeAmqpNotFoundError('exchange', destination);
+
+          const exchange = broker.getExchange(source);
+          if (!exchange) throw new FakeAmqpNotFoundError('exchange', source);
+
+          const result = broker.unbindExchange(source, destination, pattern);
+          if (!result && version <= 3.2) throw new FakeAmqpNotFoundError('binding', pattern);
+
+          return true;
+        }
+      },
+      unbindQueue(queue, source, pattern, ...args) {
+        return callBroker(check, ...args);
+
+        function check() {
+          const q = broker.getQueue(queue);
+          if (!q) throw new FakeAmqpNotFoundError('queue', queue);
+
+          const exchange = broker.getExchange(source);
+          if (!exchange) throw new FakeAmqpNotFoundError('exchange', source);
+
+          const binding = exchange.getBinding(queue, pattern);
+          if (!binding && version <= 3.2) throw new FakeAmqpNotFoundError('binding', pattern, version < 3.2);
+
+          broker.unbindQueue(queue, source, pattern);
+          return true;
+        }
+      },
+      consume(queue, onMessage, options = {}, callback) {
+        return callBroker(check, callback);
+
+        function check(callback) {
           const q = queue && broker.getQueue(queue);
           if (queue && !q) {
-            throw new FakeAmqpError(`Channel closed by server: 404 (NOT-FOUND) with message "NOT_FOUND - no queue '${queue}' in vhost '/'`, 404, true);
+            throw new FakeAmqpNotFoundError('queue', queue);
           }
 
           if (q) {
@@ -211,7 +289,9 @@ function Fake() {
             }
           }
 
-          return broker.consume(queue, onMessage && handler, {...options, channelName, prefetch});
+          const {consumerTag} = broker.consume(queue, onMessage && handler, {...options, channelName, prefetch});
+          if (callback) return callback(null, {consumerTag});
+          return {consumerTag};
         }
 
         function handler(_, msg) {
@@ -232,9 +312,11 @@ function Fake() {
       },
       ack: broker.ack,
       ackAll: broker.ackAll,
-      nack(message, ...args) {
-        return broker.nack(message, ...args);
-      },
+      ...(version >= 2.3 ? {
+        nack(message, ...args) {
+          return broker.nack(message, ...args);
+        }
+      } : undefined),
       reject: broker.reject,
       nackAll: broker.nackAll,
       prefetch(val) {
@@ -290,7 +372,6 @@ function Fake() {
         } catch (err) {
           if (err._killConnection) connection.close();
           else if (err._killChannel) closed = true;
-
           if (!poppedCb) return reject(err);
           poppedCb(err);
           return resolve();
@@ -306,9 +387,9 @@ function Fake() {
   }
 }
 
-
 function resolveOrCallback(optionalCb, err, ...args) {
   if (typeof optionalCb === 'function') optionalCb(err, ...args);
+  if (err) return Promise.reject(err);
   return Promise.resolve(...args);
 }